r159: fixed an integer overflow via voutcn@b39d757

+++ hit.c
@@ -21,7 +21,7 @@ void ma_hit_sort(size_t n, ma_hit_t *a)
 	radix_sort_hit(a, a + n);
 }
 
-void ma_hit_mark_unused(sdict_t *d, int n, const ma_hit_t *a)
+void ma_hit_mark_unused(sdict_t *d, size_t n, const ma_hit_t *a)
 {
 	size_t i;
 	for (i = 0; i < d->n_seq; ++i)
